pyproject.toml
[tool.poetry]
name = "trump-vance-debater"
version = "0.1.0"
description = ""
authors = ["Thomas Jones <thomas@ae.studio>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
streamlit = "^1.40.0"
python-dotenv = "^1.0.1"
openai = "^1.54.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
